feat: start adding the download worker
