+++ domain/application/errors/errors.go
@@ -151,4 +151,18 @@ const (
 	// AlreadyDownloadingCharm describes an error that occurs when a charm is
 	// already being downloaded.
 	AlreadyDownloadingCharm = errors.ConstError("already downloading charm")
+
+	// UnableToResolveCharm describes an error that occurs when a charm cannot
+	// be resolved.
+	UnableToResolveCharm = errors.ConstError("unable to resolve charm")
+
+	// CharmAlreadyResolved describes an error that occurs when a charm is
+	// already resolved. This means the charm for the hash already exists and
+	// can just be used.
+	CharmAlreadyResolved = errors.ConstError("charm already resolved")
+
+	// CharmNotResolved describes an error that occurs when a charm is not
+	// resolved. This means the charm for the hash does not exist and needs to
+	// be downloaded.
+	CharmNotResolved = errors.ConstError("charm not resolved")
 )
